+++ app/resources/scanned_resources/scanned_resource_wayfinder.rb
@@ -6,7 +6,6 @@ class ScannedResourceWayfinder < BaseWayfinder
   relationship_by_property :scanned_resources, property: :member_ids, model: ScannedResource
   relationship_by_property :collections, property: :member_of_collection_ids
   inverse_relationship_by_property :parents, property: :member_ids, singular: true
-  inverse_relationship_by_property :preservation_objects, property: :preserved_object_id, singular: true, model: PreservationObject
   inverse_relationship_by_property :child_tombstones, property: :parent_id, model: Tombstone
 
   def scanned_resources_count