+++ app/resources/tombstones/tombstone_change_set.rb
@@ -6,4 +6,8 @@ class TombstoneChangeSet < ChangeSet
   property :file_set_original_filename
   property :preservation_object
   property :parent_id
+
+  def preserve?
+    false
+  end
 end