+++ config/deploy/production.rb
@@ -14,6 +14,7 @@ server "lib-proc5", user: "deploy", roles: %w[worker]
 server "lib-proc6", user: "deploy", roles: %w[worker]
 server "lib-proc7", user: "deploy", roles: %w[worker]
 server "lib-proc8", user: "deploy", roles: %w[worker]
+server "lib-proc9", user: "deploy", roles: %w[worker]
 
 set :google_fixity_request_topic, "figgy-production-fixity-request"
 set :google_service_account, "figgy-preservation-production@pulibrary-figgy-storage-1.iam.gserviceaccount.com"